added reactive methods to count distance between two adresses.

+++ src/main/java/pl/marcin/project/routeservice/Client.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 public class Client {
-    Integer clientNr;
+    Long clientNr;
     Integer distance;
 
 }
